ap/services/synthesis/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
numpy
pandas
scipy
psycopg2-binary
